news_fetcher/requirements.txt
requests
pandas
tenacity
pytest
flake8
python-dotenv
psycopg2-binary